[feat][testcase] Exception handling, dependencies
Normally when running hyvr as a python module (`python -m hyvr <inifile>`),
hyvr doesn't throw exceptions, but uses sys.exit(1) for a nicer user
behaviour.
In case the default test case is running, hyvr should raise exceptions
instead, so we know exactly what went wrong.

Also, when failing to import one of the optional dependencies that are
required for some output formats, hyvr now converts the ImportError to a
warning.