[bug][trough] structure and azimuth corrections
The documentation lists 'massive' as internal structure option for
troughs, however, using this was not possible up to know, as the
structure was called 'flat' in prior versions of hyvr.
Now 'massive' is also possible (and recommended), but 'flat' will remain
for backwards compability.
'flat' in made.ini has also been changed to 'massive'

For dipping troughs, the azimuth has previously been calculated as sum
of trough azimuth and azimuth specified in the ini file. As this is not
done similarly for massive troughs, it has now been changed to only use
the azimuth value, and not the trough rotation.